+++ figures/FillSquare/FillSquare.hpp
@@ -1,14 +1,14 @@
 #ifndef FILL_SQUARE
 #define FILL_SQUARE
 
-#include <figures/FillRectangle/FillRectangle.hpp>
 #include <figures/IDraw.hpp>
+#include <figures/utils/utils.hpp>
 
 namespace top
 {
   struct FilledSquare : IDraw
   {
-    FilledSquare(p_t left_bot, int width);
+    FilledSquare(p_t left_bot, size_t width);
     p_t begin() const override;
     p_t next(p_t curr) const override;
