+++ figures/Square/Square.hpp
@@ -2,13 +2,13 @@
 #define SQUARE
 
 #include "figures/IDraw.hpp"
-#include "figures/Rectangle/Rectangle.hpp"
+#include "figures/utils/utils.hpp"
 
 namespace top
 {
   struct Square : IDraw
   {
-    Square(p_t left_bot, int width);
+    Square(p_t left_bot, size_t width);
     p_t begin() const override;
     p_t next(p_t curr) const override;
